Revised Some UI and started working on OSK
The OSK will be completely "home made" because xna acts like a douche
when it comes to windows forms etc. There will be a custom even handler
for the keyboard, that I already found from a google. I figure this is
easier for the small amount of text that will ever be inputed.