Did more Backend stuff, also added GameLibrary
The back end changes were preperations for the draw method to be
completed as well as getting a basis for the DrawShape class in and of
itself. Also added teh GameLibrary to the reference which is the library
that Michael gave us. Not sure if that will come up for you guys as well
but I added it on my end regardless. Also added it so that when you call
drawShape from the CreateShape class you have to pass it the
GraphicsDeviceManager, this allows us to draw the shapes onto the
screen.